+++ src/shared/interfaces/defillama-project-prefill.interface.ts
@@ -1,0 +1,5 @@
+import { ProjectMoreInfo } from "./project-more-info.interface";
+
+export class DefiLlamaProjectPrefill extends ProjectMoreInfo {
+  url: string;
+}