+++ src/shared/interfaces/index.ts
@@ -55,3 +55,4 @@ export * from "./organization-community.interface";
 export * from "./structured-jobpost-with-relations.interface";
 export * from "./structured-jobpost-with-applicants.interface";
 export * from "./project-with-relations.interface";
+export * from "./defillama-project-prefill.interface";