+++ src/shared/interfaces/user/index.ts
@@ -4,6 +4,7 @@ export * from "./session-object.interface";
 export * from "./github-config.interface";
 export * from "./user-role.interface";
 export * from "./user-profile.interface";
+export * from "./org-user-profile.interface";
 export * from "./user-skill.interface";
 export * from "./user-flow.interface";
 export * from "./user-email.interface";